refactor: 그래프 생성 클래스 분리
